refactor: wangyihao sync logic
